rewrite: Simplify by merging WikipediaRawReference into WikipediaReference
